feat: Plot rect's and text
The rectangles were plotted successfully.
It turned out that if you play with width and height, different
representations come out, until you find one that is aesthetically pleasing and
organized and concise.
Plotting text seems hard. It is plotted in one line and you cant easily break down the words to new lines.
This is probably why FCC states on the assignment page that you are supposed to use axes.
I found a link that explains how you can break down the text using axes:
https://gist.github.com/mbostock/7555321

This is were I have to go on.