FINISH: Plot the text, break the lines, make the text fit the tiles and do not overflow.
Project finished. All is running beautifully. Even the original solution doesnt have a better way
of showing the text. Its a horrible task because ideally you should display the text depending
dynamically on the tile size. That would be very complicated.
But it is still a very good solution and you can see everything.